PDPAcademy/Python/Study/Django/2-modul/Netflix/Netflix/requirements.txt
Django~=3.2.2
djangorestframework~=3.12.4
drf-yasg~=1.20.0
psycopg2-binary~=2.8.6
pytest==6.2.4
pytest-django==4.3.0
pytest-pythonpath==0.7.2
django-filter==23.5
